+++ CHANGES.rst
@@ -1,4 +1,4 @@
-Manager - 2.0.3 - Stable
+Manager - 2.0.4 - Stable
 =========================
 
 .. .changes
@@ -6,10 +6,18 @@ Manager - 2.0.3 - Stable
 Changes
 =======
 
+2.0.4 - Stable
+--------------
+
+| **Manager** 2.0.4 - Stable - Milestone: https://github.com/KelSolaar/Manager/issues?milestone=3&state=closed
+| **Foundations** 2.0.8 - Stable - Milestone: https://github.com/KelSolaar/Foundations/issues?milestone=6&page=1&state=closed
+
+-  Implemented better unicode support.
+
 2.0.3 - Stable
 --------------
 
-| **Manager** 2.0.3 - Stable - Milestone: https://github.com/KelSolaar/Manager/issues?milestone=2&state=closed
+**Manager** 2.0.3 - Stable - Milestone: https://github.com/KelSolaar/Manager/issues?milestone=2&state=closed
 
 -  Implemented a better dependencies resolving mechanism instead of the ranking sytem.
 -  Implemented changes from **Foundations** package.
@@ -47,7 +55,7 @@ Changes
 About
 -----
 
-| **Manager** by Thomas Mansencal – 2008 - 2012
-| Copyright© 2008 - 2012 – Thomas Mansencal – `thomas.mansencal@gmail.com <mailto:thomas.mansencal@gmail.com>`_
+| **Manager** by Thomas Mansencal – 2008 - 2013
+| Copyright© 2008 - 2013 – Thomas Mansencal – `thomas.mansencal@gmail.com <mailto:thomas.mansencal@gmail.com>`_
 | This software is released under terms of GNU GPL V3 license: http://www.gnu.org/licenses/
 | `http://www.thomasmansencal.com/ <http://www.thomasmansencal.com/>`_