+++ docs/sphinx/source/resources/pages/changes.rst
@@ -1,10 +1,18 @@
 Changes
 =======
 
+2.0.4 - Stable
+--------------
+
+| **Manager** 2.0.4 - Stable - Milestone: https://github.com/KelSolaar/Manager/issues?milestone=3&state=closed
+| **Foundations** 2.0.8 - Stable - Milestone: https://github.com/KelSolaar/Foundations/issues?milestone=6&page=1&state=closed
+
+-  Implemented better unicode support.
+
 2.0.3 - Stable
 --------------
 
-| **Manager** 2.0.3 - Stable - Milestone: https://github.com/KelSolaar/Manager/issues?milestone=2&state=closed
+**Manager** 2.0.3 - Stable - Milestone: https://github.com/KelSolaar/Manager/issues?milestone=2&state=closed
 
 -  Implemented a better dependencies resolving mechanism instead of the ranking sytem.
 -  Implemented changes from **Foundations** package.